@@ -130,9 +130,9 @@
 
 <template>
     <div>
-        <div class="card">
+        <div class="card overflow-hidden">
             <div class="card-header d-flex align-items-center justify-content-between">
-                <h5>{{this.title}}</h5>
+                <h2 class="h6 m-0">{{this.title}}</h2>
             </div>
 
 
@@ -150,18 +150,18 @@
             </div>
 
 
-            <div class="table-responsive">
+            <div class="table-responsive border-top">
                 <table v-if="ready && entry" class="table mb-0 card-bg-secondary table-borderless">
                     <tbody>
                     <tr>
-                        <td class="table-fit font-weight-bold">Time</td>
+                        <td class="table-fit text-muted">Time</td>
                         <td>
                             {{localTime(entry.created_at)}} ({{timeAgo(entry.created_at)}})
                         </td>
                     </tr>
 
                     <tr>
-                        <td class="table-fit font-weight-bold">Hostname</td>
+                        <td class="table-fit text-muted">Hostname</td>
                         <td>
                             {{entry.content.hostname}}
                         </td>
@@ -170,7 +170,7 @@
                     <slot name="table-parameters" :entry="entry"></slot>
 
                     <tr v-if="!entryPoint && job">
-                        <td class="table-fit font-weight-bold">Job</td>
+                        <td class="table-fit text-muted">Job</td>
                         <td>
                             <router-link :to="{name:'job-preview', params:{id: job.id}}" class="control-action">
                                 View Job
@@ -179,7 +179,7 @@
                     </tr>
 
                     <tr v-if="!entryPoint && request">
-                        <td class="table-fit font-weight-bold">Request</td>
+                        <td class="table-fit text-muted">Request</td>
                         <td>
                             <router-link :to="{name:'request-preview', params:{id: request.id}}" class="control-action">
                                 View Request
@@ -188,7 +188,7 @@
                     </tr>
 
                     <tr v-if="!entryPoint && command">
-                        <td class="table-fit font-weight-bold">Command</td>
+                        <td class="table-fit text-muted">Command</td>
                         <td>
                             <router-link :to="{name:'command-preview', params:{id: command.id}}" class="control-action">
                                 View Command
@@ -197,13 +197,13 @@
                     </tr>
 
                     <tr v-if="entry.tags.length">
-                        <td class="table-fit font-weight-bold">Tags</td>
+                        <td class="table-fit text-muted">Tags</td>
                         <td>
                             <router-link
                                     v-for="tag in entry.tags"
                                     :key="tag"
                                     :to="{name:resource, query: {tag: tag}}"
-                                    class="badge badge-info mr-1 font-weight-light">
+                                    class="badge badge-info mr-1">
                                 {{tag}}
                             </router-link>
                         </td>
@@ -224,7 +224,7 @@
 
             <table class="table mb-0 card-bg-secondary table-borderless">
                 <tr>
-                    <td class="table-fit font-weight-bold">ID</td>
+                    <td class="table-fit text-muted">ID</td>
 
                     <td>
                         {{entry.content.user.id}}
@@ -232,7 +232,7 @@
                 </tr>
 
                 <tr v-if="entry.content.user.name">
-                    <td class="table-fit font-weight-bold align-middle">Name</td>
+                    <td class="table-fit text-muted align-middle">Name</td>
 
                     <td class="align-middle">
                         <img :src="entry.content.user.avatar" :alt="entry.content.user.name" class="mr-2 rounded-circle" height="40" width="40" v-if="entry.content.user.avatar">
@@ -241,7 +241,7 @@
                 </tr>
 
                 <tr v-if="entry.content.user.email">
-                    <td class="table-fit font-weight-bold">Email Address</td>
+                    <td class="table-fit text-muted">Email Address</td>
 
                     <td>
                         {{entry.content.user.email}}
