@@ -43,7 +43,7 @@ export default {
 
         requestStatusClass(status) {
             if (!status) return 'danger';
-            if (status < 300) return 'secondary';
+            if (status < 300) return 'success';
             if (status < 400) return 'info';
             if (status < 500) return 'warning';
             if (status >= 500) return 'danger';
