@@ -25,7 +25,7 @@
     <preview-screen title="Batch Details" resource="batches" :id="$route.params.id" entry-point="true">
         <template slot="table-parameters" slot-scope="slotProps">
             <tr>
-                <td class="table-fit font-weight-bold">Status</td>
+                <td class="table-fit text-muted">Status</td>
                 <td>
                     <small class="badge badge-danger badge-sm" v-if="slotProps.entry.content.failedJobs > 0 && slotProps.entry.content.progress < 100">
                         Failures
@@ -40,42 +40,42 @@
             </tr>
 
             <tr v-if="slotProps.entry.content.cancelledAt">
-                <td class="table-fit font-weight-bold">Cancelled At</td>
+                <td class="table-fit text-muted">Cancelled At</td>
                 <td>
                     {{localTime(slotProps.entry.content.cancelledAt)}} ({{timeAgo(slotProps.entry.content.cancelledAt)}})
                 </td>
             </tr>
 
             <tr v-if="slotProps.entry.content.finishedAt">
-                <td class="table-fit font-weight-bold">Finished At</td>
+                <td class="table-fit text-muted">Finished At</td>
                 <td>
                     {{localTime(slotProps.entry.content.finishedAt)}} ({{timeAgo(slotProps.entry.content.finishedAt)}})
                 </td>
             </tr>
 
             <tr>
-                <td class="table-fit font-weight-bold">Batch</td>
+                <td class="table-fit text-muted">Batch</td>
                 <td>
                     {{slotProps.entry.content.name || slotProps.entry.content.id}}
                 </td>
             </tr>
 
             <tr>
-                <td class="table-fit font-weight-bold">Connection</td>
+                <td class="table-fit text-muted">Connection</td>
                 <td>
                     {{slotProps.entry.content.connection}}
                 </td>
             </tr>
 
             <tr>
-                <td class="table-fit font-weight-bold">Queue</td>
+                <td class="table-fit text-muted">Queue</td>
                 <td>
                     {{slotProps.entry.content.queue}}
                 </td>
             </tr>
 
             <tr>
-                <td class="table-fit font-weight-bold">Size</td>
+                <td class="table-fit text-muted">Size</td>
                 <td>
                     <router-link :to="{name:'jobs', query: {family_hash: slotProps.entry.family_hash}}" class="control-action">
                         {{slotProps.entry.content.totalJobs}} Jobs
@@ -84,14 +84,14 @@
             </tr>
 
             <tr>
-                <td class="table-fit font-weight-bold">Pending</td>
+                <td class="table-fit text-muted">Pending</td>
                 <td>
                     {{slotProps.entry.content.pendingJobs}}
                 </td>
             </tr>
 
             <tr>
-                <td class="table-fit font-weight-bold">Progress</td>
+                <td class="table-fit text-muted">Progress</td>
                 <td>
                     {{slotProps.entry.content.progress}}%
                 </td>
