@@ -16,14 +16,14 @@
     <preview-screen title="Command Details" resource="commands" :id="$route.params.id" entry-point="true">
         <template slot="table-parameters" slot-scope="slotProps">
             <tr>
-                <td class="table-fit font-weight-bold">Command</td>
+                <td class="table-fit text-muted">Command</td>
                 <td>
-                    {{slotProps.entry.content.command}}
+                    <code>{{slotProps.entry.content.command}}</code>
                 </td>
             </tr>
 
             <tr>
-                <td class="table-fit font-weight-bold">Exit Code</td>
+                <td class="table-fit text-muted">Exit Code</td>
                 <td>
                     {{slotProps.entry.content.exit_code}}
                 </td>
@@ -31,7 +31,7 @@
         </template>
 
         <div slot="after-attributes-card" slot-scope="slotProps">
-            <div class="card mt-5">
+            <div class="card mt-5 overflow-hidden">
                 <ul class="nav nav-pills">
                     <li class="nav-item">
                         <a class="nav-link" :class="{active: currentTab=='arguments'}" href="#" v-on:click.prevent="currentTab='arguments'">Arguments</a>
@@ -41,11 +41,10 @@
                     </li>
                 </ul>
                 <div>
-                    <div class="code-bg p-4 mb-0 text-white" v-show="currentTab=='arguments'">
-                        <vue-json-pretty :data="slotProps.entry.content.arguments"></vue-json-pretty>
-                    </div>
-                    <div class="code-bg p-4 mb-0 text-white" v-show="currentTab=='options'">
-                        <vue-json-pretty :data="slotProps.entry.content.options"></vue-json-pretty>
+                    <div class="code-bg p-4 mb-0 text-white">
+                        <copy-clipboard :data="slotProps.entry.content[currentTab]">
+                            <vue-json-pretty :data="slotProps.entry.content[currentTab]"></vue-json-pretty>
+                        </copy-clipboard>
                     </div>
                 </div>
             </div>
