@@ -39,21 +39,21 @@
     <preview-screen title="Exception Details" resource="exceptions" :id="$route.params.id">
         <template slot="table-parameters" slot-scope="slotProps">
             <tr>
-                <td class="table-fit font-weight-bold">Type</td>
+                <td class="table-fit text-muted">Type</td>
                 <td>
                     {{slotProps.entry.content.class}}
                 </td>
             </tr>
 
             <tr>
-                <td class="table-fit font-weight-bold">Location</td>
+                <td class="table-fit text-muted">Location</td>
                 <td>
                     {{slotProps.entry.content.file}}:{{slotProps.entry.content.line}}
                 </td>
             </tr>
 
             <tr>
-                <td class="table-fit font-weight-bold">Occurrences</td>
+                <td class="table-fit text-muted">Occurrences</td>
                 <td>
                     <router-link :to="{name:'exceptions', query: {family_hash: slotProps.entry.family_hash}}" class="control-action">
                         View Other Occurrences
@@ -62,21 +62,21 @@
             </tr>
 
             <tr>
-                <td class="table-fit font-weight-bold">Resolved at</td>
+                <td class="table-fit text-muted">Resolved at</td>
 
                 <td>
                     <span v-if="entry.content.resolved_at">
                         {{localTime(entry.content.resolved_at)}} ({{timeAgo(entry.content.resolved_at)}})
                     </span>
                     <span v-if="!entry.content.resolved_at">
-                        <a href="#" class="badge badge-success mr-1 font-weight-light" v-on:click.prevent="markExceptionAsResolved(entry)">Mark as resolved</a>
+                        <button class="btn btn-sm btn-success" v-on:click.prevent="markExceptionAsResolved(entry)">Mark as resolved</button>
                     </span>
                 </td>
             </tr>
         </template>
 
         <div slot="after-attributes-card" slot-scope="slotProps" class="mt-5">
-            <div class="card mt-5">
+            <div class="card mt-5 overflow-hidden">
                 <ul class="nav nav-pills">
                     <li class="nav-item">
                         <a class="nav-link" :class="{active: currentTab=='message'}" href="#" v-on:click.prevent="currentTab='message'">Message</a>
@@ -105,7 +105,9 @@
                     </code-preview>
 
                     <div class="code-bg p-4 mb-0 text-white" v-show="currentTab=='context'">
-                        <vue-json-pretty :data="slotProps.entry.content.context"></vue-json-pretty>
+                        <copy-clipboard :data="slotProps.entry.content.context">
+                            <vue-json-pretty :data="slotProps.entry.content.context"></vue-json-pretty>
+                        </copy-clipboard>
                     </div>
 
                     <stack-trace :trace="slotProps.entry.content.trace" v-show="currentTab=='trace'"></stack-trace>
