@@ -13,25 +13,25 @@
     <preview-screen title="Notification Details" resource="notifications" :id="$route.params.id">
         <template slot="table-parameters" slot-scope="slotProps">
             <tr>
-                <td class="table-fit font-weight-bold">Channel</td>
+                <td class="table-fit text-muted">Channel</td>
                 <td>
                     {{slotProps.entry.content.channel}}
                 </td>
             </tr>
 
             <tr>
-                <td class="table-fit font-weight-bold">Notification</td>
+                <td class="table-fit text-muted">Notification</td>
                 <td>
                     {{slotProps.entry.content.notification}}
 
-                    <span class="badge badge-secondary font-weight-light ml-2" v-if="slotProps.entry.content.queued">
+                    <span class="badge badge-secondary ml-2" v-if="slotProps.entry.content.queued">
                         Queued
                     </span>
                 </td>
             </tr>
 
             <tr>
-                <td class="table-fit font-weight-bold">Notifiable</td>
+                <td class="table-fit text-muted">Notifiable</td>
                 <td>
                     {{slotProps.entry.content.notifiable}}
                 </td>
