@@ -23,23 +23,23 @@
     <preview-screen title="Query Details" resource="queries" :id="$route.params.id" v-on:ready="highlightSQL()">
         <template slot="table-parameters" slot-scope="slotProps">
             <tr>
-                <td class="table-fit font-weight-bold">Connection</td>
+                <td class="table-fit text-muted">Connection</td>
                 <td>
                     {{slotProps.entry.content.connection}}
                 </td>
             </tr>
 
             <tr  v-if="slotProps.entry.content.file">
-                <td class="table-fit font-weight-bold">Location</td>
+                <td class="table-fit text-muted">Location</td>
                 <td>
                     {{slotProps.entry.content.file}}:{{slotProps.entry.content.line}}
                 </td>
             </tr>
 
             <tr>
-                <td class="table-fit font-weight-bold">Duration</td>
+                <td class="table-fit text-muted">Duration</td>
                 <td>
-                    <span class="badge badge-danger font-weight-light" v-if="slotProps.entry.content.slow">
+                    <span class="badge badge-danger" v-if="slotProps.entry.content.slow">
                         {{slotProps.entry.content.time}}ms
                     </span>
 
@@ -51,10 +51,17 @@
         </template>
 
         <div slot="after-attributes-card" slot-scope="slotProps">
-            <div class="card mt-5">
-                <div class="card-header"><h5>Query</h5></div>
-
-                <pre class="code-bg p-4 mb-0 text-white" ref="sqlcode">{{ formatSql(slotProps.entry.content.sql) }}</pre>
+            <div class="card mt-5 overflow-hidden">
+                <ul class="nav nav-pills">
+                    <li class="nav-item">
+                        <a class="nav-link active">Query</a>
+                    </li>
+                </ul>
+                <div class="code-bg p-4 mb-0 text-white">
+                    <copy-clipboard :data="formatSql(slotProps.entry.content.sql)">
+                      <pre class="code-bg text-white" ref="sqlcode">{{ formatSql(slotProps.entry.content.sql) }}</pre>
+                    </copy-clipboard>
+                </div>
             </div>
         </div>
     </preview-screen>
