@@ -13,14 +13,14 @@
     <preview-screen title="Redis Command Details" resource="redis" :id="$route.params.id">
         <template slot="table-parameters" slot-scope="slotProps">
             <tr>
-                <td class="table-fit font-weight-bold">Connection</td>
+                <td class="table-fit text-muted">Connection</td>
                 <td>
                     {{slotProps.entry.content.connection}}
                 </td>
             </tr>
 
             <tr>
-                <td class="table-fit font-weight-bold">Duration</td>
+                <td class="table-fit text-muted">Duration</td>
                 <td>
                     {{slotProps.entry.content.time}}ms
                 </td>
@@ -28,8 +28,12 @@
         </template>
 
         <div slot="after-attributes-card" slot-scope="slotProps">
-            <div class="card mt-5">
-                <div class="card-header"><h5>Command</h5></div>
+            <div class="card mt-5 overflow-hidden">
+                <ul class="nav nav-pills">
+                    <li class="nav-item">
+                        <a class="nav-link active">Command</a>
+                    </li>
+                </ul>
 
                 <pre class="code-bg p-4 mb-0 text-white">{{slotProps.entry.content.command}}</pre>
             </div>
