@@ -1,5 +1,11 @@
 <script type="text/ecmascript-6">
+    import StylesMixin from './../../mixins/entriesStyles';
+
     export default {
+        mixins: [
+            StylesMixin,
+        ],
+
         data() {
             return {
                 entry: null,
@@ -14,56 +20,60 @@
     <preview-screen title="Request Details" resource="requests" :id="$route.params.id" entry-point="true">
         <template slot="table-parameters" slot-scope="slotProps">
         <tr>
-            <td class="table-fit font-weight-bold">Method</td>
+            <td class="table-fit text-muted">Method</td>
             <td>
-                {{slotProps.entry.content.method}}
+                <span class="badge" :class="'badge-'+requestMethodClass(slotProps.entry.content.method)">
+                    {{slotProps.entry.content.method}}
+                </span>
             </td>
         </tr>
 
         <tr>
-            <td class="table-fit font-weight-bold">Controller Action</td>
+            <td class="table-fit text-muted">Controller Action</td>
             <td>
                 {{slotProps.entry.content.controller_action}}
             </td>
         </tr>
 
         <tr v-if="slotProps.entry.content.middleware">
-            <td class="table-fit font-weight-bold">Middleware</td>
+            <td class="table-fit text-muted">Middleware</td>
             <td>
                 {{slotProps.entry.content.middleware.join(", ")}}
             </td>
         </tr>
 
         <tr>
-            <td class="table-fit font-weight-bold">Path</td>
+            <td class="table-fit text-muted">Path</td>
             <td>
                 {{slotProps.entry.content.uri}}
             </td>
         </tr>
 
         <tr>
-            <td class="table-fit font-weight-bold">Status</td>
+            <td class="table-fit text-muted">Status</td>
             <td>
-                {{slotProps.entry.content.response_status}}
+                <span class="badge" :class="'badge-'+requestStatusClass(slotProps.entry.content.response_status)">
+                    {{slotProps.entry.content.response_status}}
+                </span>
             </td>
         </tr>
 
         <tr>
-            <td class="table-fit font-weight-bold">Duration</td>
+            <td class="table-fit text-muted">Duration</td>
             <td>
                 {{slotProps.entry.content.duration || '-'}} ms
             </td>
         </tr>
 
         <tr>
-            <td class="table-fit font-weight-bold">IP Address</td>
+            <td class="table-fit text-muted">IP Address</td>
             <td>
                 {{slotProps.entry.content.ip_address || '-'}}
             </td>
         </tr>
 
         <tr>
-            <td class="table-fit font-weight-bold">Memory usage</td>
+            <td class="table-fit text-muted">Memory usage</td>
             <td>
                 {{slotProps.entry.content.memory || '-'}} MB
             </td>
@@ -71,7 +81,7 @@
         </template>
 
         <div slot="after-attributes-card" slot-scope="slotProps">
-            <div class="card mt-5">
+            <div class="card mt-5 overflow-hidden">
                 <ul class="nav nav-pills">
                     <li class="nav-item">
                         <a class="nav-link" :class="{active: currentTab=='payload'}" href="#" v-on:click.prevent="currentTab='payload'">Payload</a>
@@ -87,10 +97,9 @@
                     </li>
                 </ul>
                 <div class="code-bg p-4 mb-0 text-white">
-                    <vue-json-pretty :data="slotProps.entry.content.payload" v-if="currentTab=='payload'"></vue-json-pretty>
-                    <vue-json-pretty :data="slotProps.entry.content.headers" v-if="currentTab=='headers'"></vue-json-pretty>
-                    <vue-json-pretty :data="slotProps.entry.content.session" v-if="currentTab=='session'"></vue-json-pretty>
-                    <vue-json-pretty :data="slotProps.entry.content.response" v-if="currentTab=='response'"></vue-json-pretty>
+                    <copy-clipboard :data="slotProps.entry.content[currentTab]">
+                        <vue-json-pretty :data="slotProps.entry.content[currentTab]"></vue-json-pretty>
+                    </copy-clipboard>
                 </div>
             </div>
 
