@@ -22,34 +22,15 @@ class ResumeCommand extends Command
     protected $description = 'Unpause all Telescope watchers';
 
     /**
-     * The cache repository implementation.
-     *
-     * @var \Illuminate\Contracts\Cache\Repository
-     */
-    protected $cache;
-
-    /**
-     * Create a new command instance.
+     * Execute the console command.
      *
      * @param  \Illuminate\Contracts\Cache\Repository  $cache
      * @return void
      */
-    public function __construct(CacheRepository $cache)
+    public function handle(CacheRepository $cache)
     {
-        parent::__construct();
-
-        $this->cache = $cache;
-    }
-
-    /**
-     * Execute the console command.
-     *
-     * @return void
-     */
-    public function handle()
-    {
-        if ($this->cache->get('telescope:pause-recording')) {
-            $this->cache->forget('telescope:pause-recording');
+        if ($cache->get('telescope:pause-recording')) {
+            $cache->forget('telescope:pause-recording');
         }
 
         $this->info('Telescope watchers resumed successfully.');
