@@ -8,104 +8,104 @@ use Illuminate\Http\Request;
 
 class EntryQueryOptions
 {
-    /**
-     * The batch ID that entries should belong to.
-     *
-     * @var string
-     */
-    public $batchId;
-
-    /**
-     * The tag that must belong to retrieved entries.
-     *
-     * @var string
-     */
-    public $tag;
-
-    /**
-     * The family hash that must belong to retrieved entries.
-     *
-     * @var string
-     */
-    public $familyHash;
-
-    /**
-     * The ID that all retrieved entries should be less than.
-     *
-     * @var mixed
-     */
-    public $beforeSequence;
-
-    /**
-     * The list of UUIDs of entries tor retrieve.
-     *
-     * @var mixed
-     */
-    public $uuids;
+	/**
+	 * The batch ID that entries should belong to.
+	 *
+	 * @var string
+	 */
+	public $batchId;
 
 	/**
-     * The list of UUIDs of entries tor retrieve.
-     *
-     * @var string
-     */
-    public $startTime;
+	 * The tag that must belong to retrieved entries.
+	 *
+	 * @var string
+	 */
+	public $tag;
 
 	/**
-     * The list of UUIDs of entries tor retrieve.
-     *
-     * @var string
-     */
-    public $endTime;
+	 * The family hash that must belong to retrieved entries.
+	 *
+	 * @var string
+	 */
+	public $familyHash;
 
 	/**
-     * The list of UUIDs of entries tor retrieve.
-     *
-     * @var string
-     */
-    public $aroundTime;
+	 * The ID that all retrieved entries should be less than.
+	 *
+	 * @var mixed
+	 */
+	public $beforeSequence;
 
 	/**
-     * The list of UUIDs of entries tor retrieve.
-     *
-     * @var string
-     */
-    public $ipAddress;
+	 * The list of UUIDs of entries tor retrieve.
+	 *
+	 * @var mixed
+	 */
+	public $uuids;
 
 	/**
-     * The list of UUIDs of entries tor retrieve.
-     *
-     * @var string
-     */
-    public $path;
+	 * The list of UUIDs of entries tor retrieve.
+	 *
+	 * @var string
+	 */
+	public $startTime;
 
 	/**
-     * The list of UUIDs of entries tor retrieve.
-     *
-     * @var string
-     */
-    public $method;
+	 * The list of UUIDs of entries tor retrieve.
+	 *
+	 * @var string
+	 */
+	public $endTime;
 
 	/**
-     * The list of UUIDs of entries tor retrieve.
-     *
-     * @var string
-     */
-    public $sort;
+	 * The list of UUIDs of entries tor retrieve.
+	 *
+	 * @var string
+	 */
+	public $aroundTime;
 
 	/**
-     * The list of UUIDs of entries tor retrieve.
-     *
-     * @var string
-     */
-    public $search;
+	 * The list of UUIDs of entries tor retrieve.
+	 *
+	 * @var string
+	 */
+	public $ipAddress;
+
+	/**
+	 * The list of UUIDs of entries tor retrieve.
+	 *
+	 * @var string
+	 */
+	public $path;
+
+	/**
+	 * The list of UUIDs of entries tor retrieve.
+	 *
+	 * @var string
+	 */
+	public $method;
+
+	/**
+	 * The list of UUIDs of entries tor retrieve.
+	 *
+	 * @var string
+	 */
+	public $sort;
+
+	/**
+	 * The list of UUIDs of entries tor retrieve.
+	 *
+	 * @var string
+	 */
+	public $search;
 
 
 	/**
-     * The list of UUIDs of entries tor retrieve.
-     *
-     * @var string
-     */
-    public $searchNot;
+	 * The list of UUIDs of entries tor retrieve.
+	 *
+	 * @var string
+	 */
+	public $searchNot;
 
 	/**
 	 * The number of entries to retrieve.
@@ -114,118 +114,118 @@ class EntryQueryOptions
 	 */
 	public $statusCode;
 
-    /**
-     * The number of entries to retrieve.
-     *
-     * @var int
-     */
-    public $limit = 50;
+	/**
+	 * The number of entries to retrieve.
+	 *
+	 * @var int
+	 */
+	public $limit = 50;
 
 
 
 
-    /**
-     * Create new entry query options from the incoming request.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return static
-     */
-    public static function fromRequest(Request $request)
-    {
-        return (new static)
-                ->batchId($request->batch_id)
-                ->uuids($request->uuids)
-                ->beforeSequence($request->before)
-                ->tag($request->tag)
-                ->familyHash($request->family_hash)
-                ->startTime($request->start_time, $request->use_time_zone)
-                ->endTime($request->end_time, $request->use_time_zone)
-                ->aroundTime($request->around_time)
-                ->ipAddress($request->ip_address)
-                ->path($request->path)
-                ->method($request->method)
-                ->sort($request->sort)
-                ->statusCode($request->status_code)
-                ->search($request->search)
-                ->searchNot($request->searchNot)
-                ->limit($request->take ?? 50);
-    }
+	/**
+	 * Create new entry query options from the incoming request.
+	 *
+	 * @param  \Illuminate\Http\Request  $request
+	 * @return static
+	 */
+	public static function fromRequest(Request $request)
+	{
+		return (new static)
+			->batchId($request->batch_id)
+			->uuids($request->uuids)
+			->beforeSequence($request->before)
+			->tag($request->tag)
+			->familyHash($request->family_hash)
+			->startTime($request->start_time, $request->use_time_zone)
+			->endTime($request->end_time, $request->use_time_zone)
+			->aroundTime($request->around_time)
+			->ipAddress($request->ip_address)
+			->path($request->path)
+			->method($request->method)
+			->sort($request->sort)
+			->statusCode($request->status_code)
+			->search($request->search)
+			->searchNot($request->searchNot)
+			->limit($request->take ?? 50);
+	}
 
-    /**
-     * Create new entry query options for the given batch ID.
-     *
-     * @param  string  $batchId
-     * @return static
-     */
-    public static function forBatchId(?string $batchId)
-    {
-        return (new static)->batchId($batchId);
-    }
+	/**
+	 * Create new entry query options for the given batch ID.
+	 *
+	 * @param  string  $batchId
+	 * @return static
+	 */
+	public static function forBatchId(?string $batchId)
+	{
+		return (new static)->batchId($batchId);
+	}
 
-    /**
-     * Set the batch ID for the query.
-     *
-     * @param  string  $batchId
-     * @return $this
-     */
-    public function batchId(?string $batchId)
-    {
-        $this->batchId = $batchId;
+	/**
+	 * Set the batch ID for the query.
+	 *
+	 * @param  string  $batchId
+	 * @return $this
+	 */
+	public function batchId(?string $batchId)
+	{
+		$this->batchId = $batchId;
 
-        return $this;
-    }
+		return $this;
+	}
 
-    /**
-     * Set the list of UUIDs of entries tor retrieve.
-     *
-     * @param  array  $uuids
-     * @return $this
-     */
-    public function uuids(?array $uuids)
-    {
-        $this->uuids = $uuids;
+	/**
+	 * Set the list of UUIDs of entries tor retrieve.
+	 *
+	 * @param  array  $uuids
+	 * @return $this
+	 */
+	public function uuids(?array $uuids)
+	{
+		$this->uuids = $uuids;
 
-        return $this;
-    }
+		return $this;
+	}
 
-    /**
-     * Set the ID that all retrieved entries should be less than.
-     *
-     * @param  mixed  $id
-     * @return $this
-     */
-    public function beforeSequence($id)
-    {
-        $this->beforeSequence = $id;
+	/**
+	 * Set the ID that all retrieved entries should be less than.
+	 *
+	 * @param  mixed  $id
+	 * @return $this
+	 */
+	public function beforeSequence($id)
+	{
+		$this->beforeSequence = $id;
 
-        return $this;
-    }
+		return $this;
+	}
 
-    /**
-     * Set the tag that must belong to retrieved entries.
-     *
-     * @param  string  $tag
-     * @return $this
-     */
-    public function tag(?string $tag)
-    {
-        $this->tag = $tag;
+	/**
+	 * Set the tag that must belong to retrieved entries.
+	 *
+	 * @param  string  $tag
+	 * @return $this
+	 */
+	public function tag(?string $tag)
+	{
+		$this->tag = $tag;
 
-        return $this;
-    }
+		return $this;
+	}
 
-    /**
-     * Set the family hash that must belong to retrieved entries.
-     *
-     * @param  string  $familyHash
-     * @return $this
-     */
-    public function familyHash(?string $familyHash)
-    {
-        $this->familyHash = $familyHash;
+	/**
+	 * Set the family hash that must belong to retrieved entries.
+	 *
+	 * @param  string  $familyHash
+	 * @return $this
+	 */
+	public function familyHash(?string $familyHash)
+	{
+		$this->familyHash = $familyHash;
 
-        return $this;
-    }
+		return $this;
+	}
 
 	/**
 	 * Set the family hash that must belong to retrieved entries.
@@ -355,18 +355,18 @@ class EntryQueryOptions
 		return $this;
 	}
 
-    /**
-     * Set the number of entries that should be retrieved.
-     *
-     * @param  int  $limit
-     * @return $this
-     */
-    public function limit(int $limit)
-    {
-        $this->limit = $limit;
+	/**
+	 * Set the number of entries that should be retrieved.
+	 *
+	 * @param  int  $limit
+	 * @return $this
+	 */
+	public function limit(int $limit)
+	{
+		$this->limit = $limit;
 
-        return $this;
-    }
+		return $this;
+	}
 
 	private function needConvertTimeZoneToUtc($time, $type, $useTimeZone){
 		if($useTimeZone){
