@@ -2,6 +2,7 @@
 
 namespace Laravel\Telescope\Watchers;
 
+use Exception;
 use Illuminate\Contracts\Cache\Factory as CacheFactory;
 use Laravel\Telescope\IncomingDumpEntry;
 use Laravel\Telescope\Telescope;
@@ -40,7 +41,15 @@ class DumpWatcher extends Watcher
      */
     public function register($app)
     {
-        if (! ($this->options['always'] ?? false) && ! $this->cache->get('telescope:dump-watcher')) {
+        $dumpWatcherCache = false;
+
+        try {
+            $dumpWatcherCache = $this->cache->get('telescope:dump-watcher');
+        } catch (Exception) {
+            //
+        }
+
+        if (! ($this->options['always'] ?? false) && ! $dumpWatcherCache) {
             return;
         }
 
