@@ -21,14 +21,25 @@ trait FetchesStackTrace
                 return false;
             }
 
-            return ! Str::contains($frame['file'],
-                base_path('vendor'.DIRECTORY_SEPARATOR.$this->ignoredVendorPath())
-            );
+            return ! Str::contains($frame['file'], $this->ignoredPaths());
         });
     }
 
     /**
-     * Choose the frame outside of either Telescope/Laravel or all packages.
+     * Get the file paths that should not be used by backtraces.
+     *
+     * @return array
+     */
+    protected function ignoredPaths(): array
+    {
+        return array_merge(
+            [base_path('vendor'.DIRECTORY_SEPARATOR.$this->ignoredVendorPath())],
+            $this->options['ignore_paths'] ?? []
+        );
+    }
+
+    /**
+     * Choose the frame outside of either Telescope / Laravel or all packages.
      *
      * @return string|null
      */
